requirements.txt
streamlit
streamlit-drawable-canvas
paddlepaddle
paddleocr
simple-lama-inpainting
pypdfium2
Pillow
numpy
opencv-python-headless
python-dotenv
watchdog